+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 - Add `to_qir`, `from_qir`, `to_circuit`, `inverse` method for `DMCircuit` class
 
+- Add `status` arguments as external randomness for `perfect_sampling` and `measure` methods
+
 ### Changed
 
 - Refactor circuit and dmcircuit with common methods now in `commons.py`, and merge dmcircuit.py and dmcircuit2.py, now `DMCircuit` supports MPO gate and qir